+++ sphinx/source/changelog.rst
@@ -3,6 +3,8 @@ Changelog
 
 9.2b1
 -----
+* Added the :ref:`rzxinfo.py` command (for showing the blocks in or extracting
+  the snapshots from an RZX file)
 * Added support to :ref:`tap2sna.py` for TZX block type 0x15 (direct recording)
 * :ref:`tapinfo.py` now shows info for TZX block type 0x15 (direct recording)